Part of issue #

- [x] Tests
- ~~Layman Test Client~~
- ~~Changlelog~~
- ~~Documentation~~

Optionally
- Add dependency to doc/dependencies.md

Tests on my (INDEX) desktop:
| **Dataset** | **Serial** | **Parallel** |
| --------------- | -----------: | --------------: |
| **small** | 45s | 39s |
| **all** | 5m 07s | 4m 35s |
| **ensure_all_publications** | 185s | 150s |